+++ packages/implementations/react/src/atomic/Toggle/src/index.ts
@@ -1,0 +1,5 @@
+import { Toggle } from "./Toggle";
+
+export { Toggle } from "./Toggle";
+export type { ToggleProps } from "./toggle.types";
+export default Toggle;