+++ packages/styles/src/index.ts
@@ -5,6 +5,7 @@ export * from "./components";
 export { badge } from "./packages/atomic/badge";
 export { box } from "./packages/atomic/box";
 export { button } from "./packages/atomic/button";
+export { checkbox } from "./packages/atomic/checkbox";
 export { chip } from "./packages/atomic/chip";
 export { icon } from "./packages/atomic/icon";
 export { link } from "./packages/atomic/link";
@@ -12,7 +13,7 @@ export { skeleton } from "./packages/atomic/skeleton";
 export { status } from "./packages/atomic/status";
 export { text } from "./packages/atomic/text";
 export { title } from "./packages/atomic/title";
-export { checkbox } from "./packages/atomic/checkbox";
+export { toggle } from "./packages/atomic/toggle";
 
 // types
 export type { BoxSprinkle } from "./packages/atomic/box";
@@ -26,3 +27,4 @@ export type { ButtonVariants } from "./packages/atomic/button";
 export type { CheckboxVariants } from "./packages/atomic/checkbox";
 export type { ChipVariants } from "./packages/atomic/chip";
 export type { StatusVariants } from "./packages/atomic/status";
+export type { ToggleVariants } from "./packages/atomic/toggle";