+++ packages/styles/src/packages/atomic/chip/rarui-chip.types.ts
@@ -1,5 +1,15 @@
 import { RecipeVariants } from "@vanilla-extract/recipes";
-import { styles, sprinkle } from "./rarui-chip.css";
+import { styles , paddingProperties } from "./rarui-chip.css";
+
+import { Conditions, AddDollar } from "../../../index.types";
 
 export type ChipVariants = RecipeVariants<typeof styles.chip>;
-export type ChipSprinkle = Parameters<typeof sprinkle>[0];
+
+export interface ChipSprinkle {
+  /**
+   * The padding properties are used to generate space around an chip's content area.
+   */
+  padding?:
+    | AddDollar<keyof typeof paddingProperties>
+    | Conditions<AddDollar<keyof typeof paddingProperties>>;
+}