+++ packages/styles/src/packages/atomic/icon/rarui-icon.types.ts
@@ -1,3 +1,17 @@
-import { sprinkle } from "./rarui-icon.css";
+import {
+  Conditions,
+  AddDollar,
+  StandardLonghandProperties,
+} from "../../../index.types";
+import { colorProperties } from "../../../properties";
 
-export type IconSprinkle = Parameters<typeof sprinkle>[0];
+type IconDynamicProperties = Pick<StandardLonghandProperties, "cursor">;
+
+export interface IconSprinkle extends IconDynamicProperties {
+  /**
+   * The color property sets the icon.
+   */
+  color?:
+    | AddDollar<keyof typeof colorProperties>
+    | Conditions<AddDollar<keyof typeof colorProperties>>;
+}