+++ packages/styles/src/packages/atomic/text/rarui-text.types.ts
@@ -1,3 +1,43 @@
-import { sprinkle } from "../box/rarui-box.css";
+import {
+  textFontSizeProperties,
+  textLineHeightProperties,
+} from "./rarui-text.css";
+import {
+  Conditions,
+  AddDollar,
+  StandardLonghandProperties,
+} from "../../../index.types";
+import { colorProperties, fontWeightProperties } from "../../../properties";
 
-export type TextSprinkle = Parameters<typeof sprinkle>[0];
+type TextDynamicProperties = Pick<
+  StandardLonghandProperties,
+  "textAlign" | "wordBreak"
+>;
+
+export interface TextSprinkle extends TextDynamicProperties {
+  /**
+   * The color property is used to set the color of the title.
+   */
+  color?:
+    | AddDollar<keyof typeof colorProperties>
+    | Conditions<AddDollar<keyof typeof colorProperties>>;
+  /**
+   * The color property is used to set the color of the title.
+   * @default $regular
+   */
+  fontWeight?:
+    | AddDollar<keyof typeof fontWeightProperties>
+    | Conditions<AddDollar<keyof typeof fontWeightProperties>>;
+  /**
+   * The fontSize property sets the size of the text.
+   */
+  fontSize?:
+    | AddDollar<keyof typeof textFontSizeProperties>
+    | Conditions<AddDollar<keyof typeof textFontSizeProperties>>;
+  /**
+   * The lineHeight property specifies the line height of the title.
+   */
+  lineHeight?:
+    | AddDollar<keyof typeof textLineHeightProperties>
+    | Conditions<AddDollar<keyof typeof textLineHeightProperties>>;
+}