+++ packages/implementations/react/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 `@rarui-react/components` components is a component library built with [React](https://reactjs.org).
 
+## 2024-04-25 `1.11.0`
+
+#### 🐛 Bug fixes
+
+- Fixed typing of `Box`, `Button`, `Chip`, `Link`, `Text` and `Title` atomic components. ([#27](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/27) by [@junior](https://git.rarolabs.com.br/junior))
+
+#### 🎉 New features
+
+- Added new atomic `Toggle` component. ([#27](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/27) by [@junior](https://git.rarolabs.com.br/junior))
+
 ## 2024-04-18 `1.10.0`
 
 #### 🎉 New features